+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Added
 
-- `promote` option to control whether primitive fields are gotten as primitives or nodes
+- `promote` option to control whether primitive fields are gotten as primitives or nodes, covers all primitive fields
 - `FSTView` types which dereference to singleton `FSTView` get a flag `is_one` to indicate this and can return single node on `copy/cut()` instead of slice
 - put slice to `Dict` and `MatchMapping` can accept undelimited `Dict` and `MatchMapping` source
 - `op` can be set as a global option for convenience for putting slices to `Compare`